moved business logic from service to facade
